+++ src/temporal/index.ts
@@ -1,5 +1,0 @@
-// Copyright 2026 Peter Cheng
-// Licensed under Apache-2.0
-
-export { TimeWeightConfig, TimeWeightCalculator } from "./time_aware.js";
-export type { TimeWeightConfigOptions } from "./time_aware.js";